Move utils out into separate file
